+++ src/themes.js
@@ -7,8 +7,9 @@ const themes = [
     "theme": {
       "name": "Valley Dark",
       "type": "dark",
+      "semanticHighlighting": true,
       ...editorColor('dark'),
-      ...syntaxColor('dark')
+      ...syntaxColor('dark', 'normal')
     }
   },
   {
@@ -16,10 +17,31 @@ const themes = [
     "theme": {
       "name": "Valley Light",
       "type": "light",
+      "semanticHighlighting": true,
       ...editorColor('light'),
-      ...syntaxColor('light')
+      ...syntaxColor('light', 'normal')
     }
   },
+  {
+    "id": "valley_dark_italic",
+    "theme": {
+      "name": "Valley Italic",
+      "type": "dark",
+      "semanticHighlighting": true,
+      ...editorColor('dark'),
+      ...syntaxColor('dark', 'italic')
+    }
+  },
+  {
+    "id": "valley_light_italic",
+    "theme": {
+      "name": "Valley Italic",
+      "type": "light",
+      "semanticHighlighting": true,
+      ...editorColor('light'),
+      ...syntaxColor('light', 'italic')
+    }
+  }
 ];
 
 export default themes;