+++ CHANGELOG.md
@@ -3,6 +3,26 @@ All notable changes to "Valley" will be documented in this file.
 
 ***
 
+## 0.3.0 - June 16, 2020
+`Changed`
+- Enabled "semanticHighlighting" by default
+- Simplified Sidebar-Section-Header
+- Minor text color adjustments
+
+## 0.2.0 - April 26, 2020
+`Added`
+- Italic theme introduced
+
+`Changed`
+- Removed "meta.embedded" from the puncutation style and added detailed scopes for punctuation to work correctly in PHP
+
+## 0.1.4 - April 01, 2020
+`Changed`
+- Improved contrast of whitspace rendering
+- Improved and unified assignment scopes
+- Improved punctuation color in embedded scopes
+- Unified background color for titleBar, quickInput, statusBar etc.
+
 ## 0.1.3 - March 12, 2020
 `Added`
 - Added terminal color palette with normal and bright variants